requirements.txt
python-dotenv[cli]
langchain
huggingface_hub
torch
transformers
diffusers[torch]
accelerate
streamlit
Pillow
